重新改

+++ install_nginx.sh
@@ -44,21 +44,24 @@ systemctl restart docker.service
 #拉取nginx镜像
 docker pull nginx
 #创建目录
-mkdir -p ${V_DATA}/nginx/www/default ${V_DATA}/nginx/www/vhost ${V_DATA}/nginx/logs ${V_DATA}/nginx/conf
+mkdir -p ${V_DATA}/nginx/www/default ${V_DATA}/nginx/www ${V_DATA}/nginx/var/log ${V_DATA}/nginx/etc/nginx
 #先建立nginx
 docker run --name nginx-conf -p 80:80 -d nginx
 #复制conf
-docker cp nginx-conf:/etc/nginx/nginx.conf ${V_DATA}/nginx/conf
+docker cp nginx-conf:/etc/nginx ${V_DATA}/nginx/etc/nginx
+
 #html
-docker cp nginx-conf:/usr/share/nginx/html/index.html ${V_DATA}/nginx/www/default
+# docker cp nginx-conf:/usr/share/nginx/html/index.html ${V_DATA}/nginx/www/default
 #stop nginx #删除 nginx镜像
 docker stop nginx-conf && docker rm nginx-conf
 #index.html
 # echo "<!DOCTYPE html><html><head><meta charset=\"utf-8\"><title>docker nginx</title></head><body><h1><center>docker nginx success!</center></h1><p><center>success!</center></p></body></html>" > /v_data/nginx/www/default/index.html
 #stop nginx #删除 nginx镜像 先删除
-docker stop nginx && docker rm nginx
+if [ $(docker ps -aqf "name=nginx") ]; then
+	docker stop nginx && docker rm nginx
+fi
 #启动正式的
-docker run --name nginx -p 80:80  -v ${V_DATA}/nginx/www/default:/usr/share/nginx/html -v ${V_DATA}/nginx/conf/nginx.conf:/etc/nginx/nginx.conf -v ${V_DATA}/nginx/conf/vhost:/etc/nginx/conf.d -v ${V_DATA}/nginx/logs:/var/log/nginx -d nginx 
+docker run --name nginx -p 80:80  -v ${V_DATA}/nginx/etc/nginx:/etc/nginx -v ${V_DATA}/nginx/var/log:/var/log/nginx -v ${V_DATA}/nginx/www:/www -d nginx
 #获得
 CONTAINER_ID=$(docker ps -aqf "name=nginx")
